feat containers/string: move string container's related headers to containers/dynamic/string directory
