+++ main.c
@@ -1,11 +1,11 @@
 #include <utils/format_print.h>
 #include <utils/format.h>
 
-#include <containers/string.h>
-#include <containers/string_impl.h>
+#include <containers/dynamic/string/string.h>
+#include <containers/dynamic/string/string_impl.h>
 
 #include <tests/containers/array/array.h>
-#include <tests/containers/string/string.h>
+#include <tests/containers/dynamic/string/string.h>
 #include <tests/utils/result/result.h>
 #include <tests/fs/file/encoding.h>
 #include <tests/network/socket/socket.h>
