+++ src/containers/CMakeLists.txt
@@ -1,8 +1,8 @@
 target_sources(${PROJECT_NAME}
     PRIVATE
+        common.h
         vector.h
         array.h
-        string.h
-        string_impl.h
-        string_impl.c
 )
+
+add_subdirectory(dynamic)
