+++ src/containers/dynamic/CMakeLists.txt
@@ -2,3 +2,5 @@ target_sources(${PROJECT_NAME}
     PRIVATE
         common.h
 )
+
+add_subdirectory(string)
