+++ src/containers/dynamic/common.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include <containers/common.h>
+#include <memory/memory.h>
+
+#define NAMESPACE_CONTAINERS_DYNAMIC(NAME) NAMESPACE_CONTAINERS(CONCAT3(dynamic, __, NAME))
