+++ src/containers/dynamic/string/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(${PROJECT_NAME}
+    PRIVATE
+        string.h
+        string_impl.h
+        string_impl.c
+)
