+++ src/containers/dynamic/string/string_impl.c
