+++ src/network/address_info/address_info.h
@@ -3,7 +3,7 @@
 #include <network/socket/socket.h>
 #include <network/common.h>
 
-#include <containers/string_impl.h>
+#include <containers/dynamic/string/string_impl.h>
 
 #if defined(WIN32)
     #include <ws2def.h>
