+++ src/tests/containers/CMakeLists.txt
@@ -1,2 +1,2 @@
 add_subdirectory(array)
-add_subdirectory(string)
+add_subdirectory(dynamic)
