+++ src/tests/containers/dynamic/string/string.c
@@ -1,9 +1,9 @@
 #include "string.h"
 
-#include <containers/string.h>
+#include <containers/dynamic/string/string.h>
 #include <utils/string/helpers.h>
 
-#include <containers/string_impl.h>
+#include <containers/dynamic/string/string_impl.h>
 #include <utils/string/comparators_impl.h>
 
 IMPLEMENT_TYPE_TESTS(string) {
