+++ src/utils/string/comparators_impl.c
@@ -1,5 +1,5 @@
 #include "comparators_impl.h"
 
-#include <containers/string_impl.h>
+#include <containers/dynamic/string/string_impl.h>
 
 IMPLEMENT_COMPARE_STRING_WITH_BUFFER();
