+++ src/utils/string/comparators_impl.h
@@ -1,6 +1,6 @@
 #pragma once
 
 #include "helpers.h"
-#include <containers/string_impl.h>
+#include <containers/dynamic/string/string_impl.h>
 
 DECLARE_COMPARE_STRING_WITH_BUFFER();
